requirements.txt
Django==2.2.4
oneagent-sdk==1.2.1
